Change admin account to read `REVIEW_ADMIN` env variable
Refactored initialization code in `set_initial_admin_password`
to check for the presence of the `REVIEW_ADMIN` environment
variable `admin:password`. If available, parses the variable to
obtain the username and password for the initial administrator
account. Returns an error if the REVIEW_ADMIN variable is invalid
or unavailable. This change eliminates hardcoded credentials,
improving security and flexibility.

Close: #38